Revert previous changes
