--- plugins/awscdk/cdk.go
@@ -1,25 +1,0 @@
-package awscdk
-
-import (
-	"github.com/1Password/shell-plugins/sdk"
-	"github.com/1Password/shell-plugins/sdk/needsauth"
-	"github.com/1Password/shell-plugins/sdk/schema"
-	"github.com/1Password/shell-plugins/sdk/schema/credname"
-)
-
-func AWSCDKCLI() schema.Executable {
-	return schema.Executable{
-		Name:    "AWS CDK Toolkit",
-		Runs:    []string{"cdk"},
-		DocsURL: sdk.URL("https://docs.aws.amazon.com/cdk/v2/guide/cli.html"),
-		NeedsAuth: needsauth.IfAll(
-			needsauth.NotForHelpOrVersion(),
-			needsauth.NotWithoutArgs(),
-		),
-		Uses: []schema.CredentialUsage{
-			{
-				Name: credname.AccessKey,
-			},
-		},
-	}
-}
